refactor: 리뷰 반영 - OrderableShopSearchResult 분리 - 개행, 불필요 코드 제거 - 파라미터, 변수, 메서드 이름 수정
